Final changes comming from TFS (Visual Studio Online)
All changset described in wod document and readme file. Nico Macedo has
access on Visual Studio Online to see them all.

+++ MeliHackComparer/MeliHackComparerWeb/Models/User.cs
@@ -29,9 +29,9 @@ namespace MeliSample.Models
 
     public class Ratings
     {
-        public int positive { get; set; }
-        public int negative { get; set; }
-        public int neutral { get; set; }
+        public decimal positive { get; set; }
+        public decimal negative { get; set; }
+        public decimal neutral { get; set; }
     }
 
     public class Transactions
